+++ tests/frontend/edu/brown/costmodel/TestTimeIntervalCostModel.java
@@ -256,7 +256,7 @@ public class TestTimeIntervalCostModel extends BaseTestCase {
         } // FOR
         final Map<Integer, Double> probs = new HashMap<Integer, Double>();
         for (int i = 0; i < NUM_PARTITIONS; i++) {
-            int cnt = h.get(i, 0);
+            long cnt = h.get(i, 0);
             probs.put(i, cnt / 100.0d);
         } // FOR
